financial_analyst_agent/requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform
google-auth
pytest
